app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
celery
redis
pillow
requests
